+++ lib/roo_on_rails.rb
@@ -6,4 +6,6 @@ end
 if defined?(Rails)
   require 'dotenv/rails-now'
   require 'roo_on_rails/railtie'
+  require 'roo_on_rails/railties/new_relic'
+  require 'roo_on_rails/railties/http'
 end