+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+Features: 
+
+ - Add Rack::Timeout
+ - Configures SSL for rails
+ - Add Rack::Deflater 
+
 # v1.1.0 (2017-02-20)
 
 Features: