+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# Unreleased
+
+Features:
+
+ - Add Rack::Timeout
+ - Add Rack::SslEnforcer
+ - Add Rack::Deflater
+
 # v1.1.0 (2017-02-20)
 
 Features:
@@ -15,4 +23,3 @@ Bug fixes:
 Features:
 
 - Automatic New Relic configuration, with safeguards
-