+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 Features:
 
-- Sets database statement timeout to 200ms by default (#10).
+- Sets database statement timeout to 200ms by default (#13).
 
 # v1.2.0 (2017-03-21)